feat: agregar redux de usuario y reorganizar estructura del proyecto
- Se agregó el slice y reducer de usuario.
- Se configuró el store para incluir el nuevo slice.
- Se reorganizó la estructura del proyecto creando nuevas carpetas para mejorar la organización.
