+++ src/main.jsx
@@ -3,8 +3,8 @@ import App from './App.jsx';
 import { BrowserRouter } from 'react-router-dom';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import { Provider } from 'react-redux';
-import store from './store.jsx';
-import { fetchProductos } from './ProductosSlice.jsx';
+import store from './store/store.jsx';
+import { fetchProductos } from './store/ProductosSlice.jsx';
 
 store.dispatch(fetchProductos());
