--- src/store.jsx
@@ -1,11 +1,0 @@
-import { configureStore } from "@reduxjs/toolkit";
-import  useReducer  from "./ProductosSlice.jsx";    
-
-
-
-
-export default configureStore({
-    reducer: {
-        productos: useReducer,    
-    }
-});
