+++ src/store/store.jsx
@@ -1,0 +1,12 @@
+import { configureStore } from "@reduxjs/toolkit";
+import  ProductosReducer  from "./ProductosSlice.jsx";    
+import UsersReducer from "./UsersSlice.jsx";
+
+
+
+export default configureStore({
+    reducer: {
+        productos: ProductosReducer,
+        users: UsersReducer    
+    }
+});
